# docs/openapi.yaml
basePath: /api/v1
definitions:
...
  dto.AdminRefundReconciliationRequest:
    properties:
      action:
        enum:
        - deduct_remaining_coins
        - accept_consumed_coins
        - revoke_premium_days
        - revoke_refunded_subscription
        - retain_entitlement
        - mark_refund_rejected
        - complete_manual_review
        type: string
      manual_review_confirmed:
        type: boolean
      note:
        maxLength: 1000
        minLength: 3
        type: string
      premium_days_to_revoke:
        maximum: 3650
        minimum: 0
        type: integer
      provider_rejection_confirmed:
        type: boolean
    required:
    - action
    - note
    type: object
  dto.AdminRefundRequest:
    properties:
      amount:
        type: integer
      reason:
        maxLength: 255
        type: string
    required:
    - amount
    - reason
    type: object
...
  dto.MidtransRefundDetail:
    properties:
      bank_confirmed_at:
        type: string
      created_at:
        type: string
      reason:
        type: string
      refund_amount:
        type: string
      refund_chargeback_id:
        type: string
      refund_key:
        type: string
      refund_method:
        type: string
    type: object
  dto.MidtransWebhookRequest:
    properties:
      bank_confirmed_at:
        type: string
      fraud_status:
        type: string
      gross_amount:
        type: string
      order_id:
        type: string
      payment_type:
        type: string
      reason:
        type: string
      refund_amount:
        type: string
      refund_chargeback_id:
        type: string
      refund_key:
        type: string
      refund_method:
        type: string
      refunds:
        items:
          $ref: '#/definitions/dto.MidtransRefundDetail'
        type: array
      settlement_time:
        type: string
      signature_key:
        type: string
      status_code:
        type: string
      transaction_id:
        type: string
      transaction_status:
        type: string
      transaction_time:
        type: string
    type: object
...
  dto.Response:
    properties:
      code:
        type: string
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@ruangtenang.id
    name: API Support
  description: API untuk aplikasi Ruang Tenang - Platform Kesehatan Mental
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Ruang Tenang API
  version: "1.0"
paths:
...
  /admin/billing/transactions/{orderId}/refund-reconciliation:
    post:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: orderId
        required: true
        type: string
      - description: Reconciliation action and operator note
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AdminRefundReconciliationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Resolve a refund case requiring operator reconciliation
      tags:
      - Billing
  /admin/billing/transactions/{orderId}/refund-status/sync:
    post:
      parameters:
      - description: Order ID
        in: path
        name: orderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/dto.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Synchronize a transaction and refund details from Midtrans
      tags:
      - Billing
  /admin/billing/transactions/{orderId}/refunds:
    post:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: orderId
        required: true
        type: string
      - description: Refund amount and reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AdminRefundRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/dto.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Request a Midtrans refund for a paid transaction
      tags:
      - Billing
...
  /billing/webhooks/midtrans:
    post:
      consumes:
      - application/json
      parameters:
      - description: Midtrans notification
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.MidtransWebhookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Receive a verified Midtrans payment notification
      tags:
      - Billing
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"